Take user to requested section even if no section is set in session
If no section is set in the session (i.e. it's the first time the user
is accessing this course in this session), the section number passed
in the URL was being ignored and the session's section number set to 0.
This resulted in users being taken to the course homepage even if they
were attempting to get to a specific section; clicking the link again
would take them to the section.